Refactored code, added documentatipon
