+++ docker_pull_and_run.yml
@@ -75,6 +75,7 @@
           MONGO_PORT: "27017"
           MONGO_DB_NAME: "{{ MONGO_DB_NAME }}"
           SCHEMA_PATH: "{{ SCHEMA_PATH | default(omit) }}"
+          ROOT_PATH: "/api"
         networks:
           - name: data_board_network
